requirements.txt
django==4.2.1
djangorestframework==3.14.0
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
drf-extra-fields==3.5.0